Second refactor/cleanup pass
